[2.2.1][Feat] - ScopeTree selection fix
